Refactor: [FE] expense sub page 에서 내려주던 props를 하위 컴포넌트에서 관리토록 이관
